homework about modules first try
