+++ src/calcs_counters_common.rs
@@ -1,0 +1,8 @@
+mod counters {
+    mod signed_counter;
+    mod unsigned_counter;
+}
+mod calcs {
+    mod pair_calc;
+    mod vec3_calc;
+}
